+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Added `DropDownOptions.sortAfterSearch` option to enable sorting. `DropDownOptions.sortDelegate` is still optional, as now `DropDownItem.compareTo` can be used by default.
 * Moved default item builder to `DropDownItem.build`.
 * Added `DropDownItemBuilder` interface to allow the `T` of `DropDownItem<T>` to determine the widget displayed in the drop down list.
+* Moved search logic to `DropDownItem.satisfiesSearch` which is now called from `DropDownList.search`.
 
 ## 0.0.4